requirements.txt
django
gunicorn
django-cors-headers
django-heroku
djangorestframework
numpy
Pillow
scikit-image
scikit-learn==0.21.3